pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "windows-service-process-monitoring-agent"
version = "0.1.0"
description = "Production-ready Windows process monitoring agent with API and dashboard."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Pratik Barbudhe" }]
dependencies = [
    "psutil>=5.9.0",
    "pywin32>=305",
    "streamlit>=1.28.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
monitor-agent = "main:main"

[tool.setuptools]
packages = ["app", "api", "dashboard", "services", "cli"]
